[a0c02d0f] fix(panel): move hook test wrappers to components and rename providers.tsx to unshadow barrel
