+++ panel/src/hooks/__tests__/use-page-refresh.test.tsx
@@ -1,17 +1,11 @@
 import { describe, it, expect, vi } from "vitest";
 import { act, renderHook, waitFor } from "@testing-library/react";
-import type { ReactNode } from "react";
 import { usePageRefresh } from "../use-page-refresh";
 import { usePageRefresh as usePageRefreshPublic } from "@/hooks";
-import { PageRefreshProvider } from "@/components/providers/page-refresh-provider";
-
-function wrapper({ children }: { children: ReactNode }) {
-  return <PageRefreshProvider>{children}</PageRefreshProvider>;
-}
-
-function disabledWrapper({ children }: { children: ReactNode }) {
-  return <PageRefreshProvider disabled>{children}</PageRefreshProvider>;
-}
+import {
+  PageRefreshWrapper,
+  DisabledPageRefreshWrapper,
+} from "@/components/providers/__tests__/test-utils";
 
 describe("usePageRefresh", () => {
   it("throws when used outside a PageRefreshProvider", () => {
@@ -21,21 +15,25 @@ describe("usePageRefresh", () => {
   });
 
   it("returns disabled=false and loading=false by default", () => {
-    const { result } = renderHook(() => usePageRefresh(), { wrapper });
+    const { result } = renderHook(() => usePageRefresh(), {
+      wrapper: PageRefreshWrapper,
+    });
     expect(result.current.disabled).toBe(false);
     expect(result.current.loading).toBe(false);
   });
 
   it("reflects the provider disabled prop", () => {
     const { result } = renderHook(() => usePageRefresh(), {
-      wrapper: disabledWrapper,
+      wrapper: DisabledPageRefreshWrapper,
     });
     expect(result.current.disabled).toBe(true);
   });
 
   it("registers and invokes callbacks on refresh", async () => {
     const cb = vi.fn();
-    const { result } = renderHook(() => usePageRefresh(), { wrapper });
+    const { result } = renderHook(() => usePageRefresh(), {
+      wrapper: PageRefreshWrapper,
+    });
 
     act(() => {
       result.current.register(cb);
@@ -50,7 +48,9 @@ describe("usePageRefresh", () => {
 
   it("unregisters callbacks", async () => {
     const cb = vi.fn();
-    const { result } = renderHook(() => usePageRefresh(), { wrapper });
+    const { result } = renderHook(() => usePageRefresh(), {
+      wrapper: PageRefreshWrapper,
+    });
 
     act(() => {
       result.current.register(cb);
@@ -71,7 +71,9 @@ describe("usePageRefresh", () => {
     });
     const cb = vi.fn(() => deferred);
 
-    const { result } = renderHook(() => usePageRefresh(), { wrapper });
+    const { result } = renderHook(() => usePageRefresh(), {
+      wrapper: PageRefreshWrapper,
+    });
     act(() => result.current.register(cb));
 
     let refreshPromise: Promise<void>;
@@ -93,7 +95,7 @@ describe("usePageRefresh", () => {
   it("does not run callbacks while disabled", async () => {
     const cb = vi.fn();
     const { result } = renderHook(() => usePageRefresh(), {
-      wrapper: disabledWrapper,
+      wrapper: DisabledPageRefreshWrapper,
     });
 
     act(() => result.current.register(cb));
@@ -116,7 +118,9 @@ describe("usePageRefresh", () => {
     });
     const cb = vi.fn(() => deferred);
 
-    const { result } = renderHook(() => usePageRefresh(), { wrapper });
+    const { result } = renderHook(() => usePageRefresh(), {
+      wrapper: PageRefreshWrapper,
+    });
     act(() => result.current.register(cb));
 
     let first: Promise<void>;
